requirements.txt
os
numpy
tensorflow
matplotlib
seaborn
scikit-learn
streamlit
Pillow
opencv-python
